move nojobs into its own component
