+++ src/components/SubmitJob/index.js
@@ -195,7 +195,6 @@ const SumbitJobPage = ({createJobPost, currentUser, logIn}) => {
                 </div>
                 :
                 <div className='mb-4'>
-                    <h3 className='mb-4'>You need to log in to post a job</h3>
                     <SignIn logIn={logIn}></SignIn>
                 </div>
             }
